backend/pyproject.toml
[project]
name = "shuren-backend"
version = "0.1.0"
description = "Shuren AI Fitness Backend API"
authors = [
    {name = "Shuren Team"}
]
readme = "README.md"
package-mode = false
requires-python = "^3.11"
dependencies = [
    "fastapi (>=0.128.0,<0.129.0)",
    "uvicorn[standard] (>=0.40.0,<0.41.0)",
    "sqlalchemy (>=2.0.46,<3.0.0)",
    "asyncpg (>=0.31.0,<0.32.0)",
    "alembic (>=1.18.2,<2.0.0)",
    "pydantic (>=2.12.5,<3.0.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "google-auth (>=2.48.0,<3.0.0)",
    "aiofiles (>=25.1.0,<26.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "email-validator (>=2.3.0,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=9.0.2,<10.0.0)",
    "pytest-asyncio (>=1.3.0,<2.0.0)",
    "pytest-cov (>=7.0.0,<8.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "aiosqlite (>=0.22.1,<0.23.0)",
    "hypothesis (>=6.151.4,<7.0.0)"
]

[tool.pytest.ini_options]
# Async test configuration
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

# Test discovery patterns
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

# Output and logging
addopts = [
    "-v",                          # Verbose output
    "--strict-markers",            # Strict marker validation
    "--tb=short",                  # Short traceback format
    "--disable-warnings",          # Disable warnings in output
    "-ra",                         # Show summary of all test outcomes
]

# Markers for test categorization
markers = [
    "unit: Unit tests for individual components",
    "integration: Integration tests requiring database",
    "property: Property-based tests using Hypothesis",
    "slow: Tests that take significant time to run",
    "auth: Authentication and authorization tests",
    "onboarding: Onboarding flow tests",
    "profile: User profile tests",
]

# Coverage configuration
[tool.coverage.run]
source = ["app"]
omit = [
    "*/tests/*",
    "*/migrations/*",
    "*/__pycache__/*",
    "*/venv/*",
    "*/.venv/*",
]
branch = true

[tool.coverage.report]
precision = 2
show_missing = true
skip_covered = false
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]

[tool.coverage.html]
directory = "htmlcov"
